+++ CHANGELOG.md
@@ -34,7 +34,7 @@
 - add tiktokkid to meme pack
 - remove support python 3.9
 - rework if else to match/case
-- add utils.formet_file_size
+- add utils.format_file_size
 - add utils.is_url
 - add utils.get_iso_time
 - add utils.safe_getattr
@@ -46,7 +46,6 @@
 - add utils.get_network_interfaces
 - add utils.get_disk_usage
 - add utils.get_ip_address
-- add utils.get_cpu_temperature
 - add utils.get_args_bool
 - add utils.get_args_int
 - fix pack_url